chore: lcov collect and ci process changed for code coverage
