+++ .github/workflows/ci.yml
@@ -62,10 +62,16 @@ jobs:
       - uses: ./.github/actions/setup-node
       - name: Test
         run: pnpm run test
-      - name: Codecov
-        uses: codecov/codecov-action@v3
-        env:
-          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      - name: Setup LCOV
+        uses: hrishikesh-kadam/setup-lcov@v1
+      - name: Report code coverage
+        uses: zgosalvez/github-actions-report-lcov@v3
+        with:
+          coverage-files: coverage/*.lcov.info
+          minimum-coverage: 90
+          update-comment: true
+          artifact-name: templ-code-coverage-report
+          github-token: ${{ secrets.GITHUB_TOKEN }}
 
   test_cli:
     runs-on: ${{ matrix.os }}
@@ -83,27 +89,3 @@ jobs:
           node-version: ${{ matrix.node_version }}
       - name: Test CLI
         run: pnpm run test:cli
-
-  # coverage_report:
-  #   needs: test
-  #   runs-on: ubuntu-latest
-  #   needs: ['clean', 'test']
-  #   steps:
-  #   - name: Checkout code
-  #     uses: actions/checkout@v3
-  #   - uses: ./.github/actions/setup-node
-  #   - name: Install
-  #     run: pnpm i --no-frozen-lockfile
-  #   - name: Coverage lcov
-  #     run: pnpm run coverage:lcov
-  #   - name: Setup LCOV
-  #     uses: hrishikesh-kadam/setup-lcov@v1
-  #   - name: Report code coverage
-  #     uses: zgosalvez/github-actions-report-lcov@v3
-  #     with:
-  #       coverage-files: coverage/*.lcov.info
-  #       minimum-coverage: 90
-  #       artifact-name: templ-code-coverage-report
-  #       github-token: ${{ secrets.GITHUB_TOKEN }}
-  #       working-directory: ./
-  #       update-comment: true
