+++ scripts/package.json
@@ -17,7 +17,7 @@
   "scripts": {
     "clean": "wireit",
     "clean:nm": "wireit",
-    "coverage:lcov": "wireit",
+    "lcov:collect": "wireit",
     "create:pkg": "wireit",
     "eslint:fix": "eslint . --fix",
     "eslint": "wireit",
@@ -30,6 +30,7 @@
     "@templ/logger": "workspace:^",
     "@templ/utils": "workspace:^",
     "fast-glob": "^3.3.1",
+    "lcov-result-merger": "^4.1.0",
     "markdown-table": "^3.0.3",
     "pretty-bytes": "^6.1.1",
     "rollup": "^3.29.2",
@@ -40,10 +41,8 @@
   "devDependencies": {
     "@antfu/eslint-config": "1.0.0-beta.2",
     "@templ/docs": "workspace:^",
-    "@types/lcov-parse": "^1.0.0",
     "@types/node": "^20.6.3",
     "eslint": "^8.50.0",
-    "lcov-parse": "^1.0.0",
     "markdownlint": "^0.30.0",
     "ora": "^7.0.1",
     "rimraf": "^5.0.1",
@@ -72,10 +71,10 @@
         }
       ]
     },
-    "coverage:lcov": {
-      "command": "node --loader tsx ./src/coverage-lcov.ts",
+    "lcov:collect": {
+      "command": "node --loader tsx ./src/lcov-collect.ts",
       "files": [
-        "src/coverage-lcov.ts"
+        "src/lcov-collect.ts"
       ],
       "dependencies": [
         {
